requirements.txt
asgiref==3.8.1
Django==5.0.7
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
drf-yasg==1.21.7
djangorestframework==3.15.1
django-cors-headers==4.4.0
Jinja2==3.1.4
pillow==10.4.0